Temporary Testing
It gives a circle (Later vertices) that is red and when clicked on can be changed to other options which should be defined as MenuItems. These MenuItems should be stored in an Array made by the MenuItemArray class which is bugged atm. So for now it just has 1 option to make it blue...