Elliot's Vertex class
Everything used to be in the testing code but it makes more sense to add it to the vertex. I didn't understand the last vertex class so I made my own. We can later use the Adjacency Matrix class to create the lines between the vertices (Using getX + getY from circle class which is it's superclass)